+++ src/utils/trip.js
@@ -2,22 +2,6 @@ import dayjs from 'dayjs';
 
 const DATE_FORMAT = 'DD/MM/YYYY HH:mm';
 const DATE_FORMAT_TIME = 'HH:mm';
-const getRandomPositiveInteger = (a, b) => {
-  if (a < 0 || b < 0) {
-    return NaN;
-  }
-
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  const result = Math.random() * (upper - lower + 1) + lower;
-
-  return Math.floor(result);
-};
-
-function getRandomArrayElement(items) {
-  return items[Math.floor(Math.random() * items.length)];
-}
 
 function humanizeTaskDueDate(dueDate) {
   return dueDate ? dayjs(dueDate).format(DATE_FORMAT) : '';
@@ -31,10 +15,6 @@ function humanizeDate(date, formatData) {
   return date ? dayjs(date).format(formatData) : '';
 }
 
-// function isOptionChecked(repeating) {
-//   return Object.values(repeating).some(Boolean);
-// }
-
 function makingOffersByType (items, typeOfOffer, checkedTrip) {
   for (let i = 0; i < items.length; i++) {
     if (items[i].type === typeOfOffer) {
@@ -45,4 +25,4 @@ function makingOffersByType (items, typeOfOffer, checkedTrip) {
 }
 
 
-export {makingOffersByType ,getRandomArrayElement, getRandomPositiveInteger, humanizeTaskDueDate, humanizeTaskDueTime, humanizeDate};
+export {makingOffersByType, humanizeTaskDueDate, humanizeTaskDueTime, humanizeDate};
